% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{compute_multi_reg}
\alias{compute_multi_reg}
\title{compute_multi_reg}
\usage{
compute_multi_reg(..., df, x_names, y_names, with_ic = T)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
compute_multi_reg
}
