% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{compute_rda}
\alias{compute_rda}
\title{compute_rda}
\usage{
compute_rda(df, env_cols, significance_tests = F, plt_model, width, height)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
compute_rda
}
