% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{create_subfolders}
\alias{create_subfolders}
\title{create_subfolders}
\usage{
create_subfolders(x, folder_path, indent = 1)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
create_subfolders
}
