% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{density_overlap_2D}
\alias{density_overlap_2D}
\title{density_overlap_2D}
\usage{
density_overlap_2D(
  scores_1,
  scores_2,
  name_1,
  name_2,
  n_env_cells_per_dim = 100,
  xlims,
  ylims,
  plt_densities = F
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
density_overlap_2D
}
