% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{distancesAndOverlaps}
\alias{distancesAndOverlaps}
\title{distancesAndOverlaps}
\usage{
distancesAndOverlaps(
  ID,
  env_df,
  class_defs,
  overlap_metric = "D",
  n_env_cells_per_dim = 100,
  plot_niches = F,
  plot_overlap = F,
  plot_dists = F,
  width = 1200,
  height = 800,
  save_folder = NULL,
  verb = 0
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
distancesAndOverlaps
}
