% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{downsampleForRDA}
\alias{downsampleForRDA}
\title{downsampleForRDA}
\usage{
downsampleForRDA(lc_env_df, save_folder = NULL)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
downsampleForRDA
}
