% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{extractDataGrid}
\alias{extractDataGrid}
\title{extractDataGrid}
\usage{
extractDataGrid(
  grid,
  lc_env_folder,
  clim_flnames,
  lc_flname,
  elev_flname,
  all_lc_codes,
  del_classes = NULL,
  merge_classes,
  class_defs,
  plt_aggregated = F,
  width = 1200,
  height = 800,
  save_folder = NULL,
  n_parallel = 1
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
extractDataGrid
}
