% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{extractDataPSA}
\alias{extractDataPSA}
\title{extractDataPSA}
\usage{
extractDataPSA(
  poly,
  landcover,
  elevation,
  climate,
  save_folder,
  del_classes = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
extractDataPSA
}
