% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{extractDataPSAs}
\alias{extractDataPSAs}
\title{extractDataPSAs}
\usage{
extractDataPSAs(
  polys,
  lc_env_folder,
  clim_flnames,
  lc_flname,
  elev_flname,
  save_folder = NULL,
  del_classes = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
extractDataPSAs
}
