% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{interpolate_ages_per_site}
\alias{interpolate_ages_per_site}
\title{modernShares}
\usage{
interpolate_ages_per_site(
  df,
  meta_cols,
  fc = 1/500,
  dt = 500,
  k = 5,
  int.method = "linear",
  appliedFilter = "gauss"
)
}
\arguments{
\item{veg_cover}{save_folder}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
...
}
