% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{interpVegTS}
\alias{interpVegTS}
\title{modernShares}
\usage{
interpVegTS(veg_cover, save_folder = NULL)
}
\arguments{
\item{veg_cover}{save_folder}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
can have one or multiple IDs in columns Dataset_ID. If multiple, Interpolation is done seperately and the result is joined.
}
