% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{k_fold_cv}
\alias{k_fold_cv}
\title{k_fold_cv}
\usage{
k_fold_cv(
  ...,
  data,
  k,
  x_names,
  y_names,
  analysis_function,
  set_seed = F,
  seed_int = NULL,
  prnt = F,
  tune = F,
  average = "num",
  verb = 1
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
k_fold_cv
}
