% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\docType{package}
\name{LandCoverChange-package}
\alias{LandCoverChange-package}
\title{Simulation of Landcover change over time, based on sedimentary pollen record time series}
\description{
....
}
\author{
Peter Ewald and Simeon Lisovski.
}
