% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{modernLCShares}
\alias{modernLCShares}
\title{modernLCShares}
\usage{
modernLCShares(
  polygons,
  resolution,
  class_defs,
  merge_classes = NULL,
  save_folder = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
modernLCShares
}
