% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{modernVegShares}
\alias{modernVegShares}
\title{modernShares}
\usage{
modernVegShares(
  veg_cover,
  time_span_modern,
  meta_cols_veg = c("Dataset_ID", "Age", "Intersects_regMap", "Intersects_calibMap"),
  save_folder = NULL
)
}
\arguments{
\item{veg_cover}{save_folder}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
...
}
