% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{multi_reg_analysis}
\alias{multi_reg_analysis}
\title{multi_reg_analysis}
\usage{
multi_reg_analysis(
  ...,
  df_train,
  df_test,
  x_names,
  y_names,
  prnt = F,
  tune = F,
  verb = 1
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
multi_reg_analysis
}
