% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{multi_reg_metrics}
\alias{multi_reg_metrics}
\title{multi_reg_metrics}
\usage{
multi_reg_metrics(..., predicted, actual, per_class = F, metric_name = "MAE")
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
multi_reg_metrics
}
