% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{plot_distances}
\alias{plot_distances}
\title{plot_distances}
\usage{
plot_distances(
  distance_matrix,
  classes,
  class_colors,
  save_folder = NULL,
  width = 1200,
  height = 800,
  scale_to_max = T
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
plot_distances
}
