% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{plot_env_space}
\alias{plot_env_space}
\title{plot_env_space}
\usage{
plot_env_space(
  df,
  class_defs,
  rda_model,
  variance_df = NULL,
  sig_df = NULL,
  p_val = NULL,
  title = "Environmental Space",
  RDA1 = "RDA1",
  RDA2 = "RDA2",
  xlims = NULL,
  ylims = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
plot_env_space
}
