% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{plotTS}
\alias{plotTS}
\title{plotVegSite}
\usage{
plotTS(
  data,
  cols,
  age_name,
  round_ages = T,
  age_steps = 1000,
  color,
  lines_var = NULL,
  lwd = 1,
  alpha = 1,
  share_to_percent = F,
  title = "",
  subtitle = ""
)
}
\arguments{
\item{veg_cover}{save_folder}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
...
}
