% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{plotVegSite}
\alias{plotVegSite}
\title{plotVegSite}
\usage{
plotVegSite(
  vegCover1,
  name1,
  vegCover2 = NULL,
  name2 = NULL,
  ID = NULL,
  meta_cols = c("Dataset_ID", "Age", "Intersects_regMap", "Intersects_calibMap"),
  save_folder = NULL,
  n_taxa = NULL
)
}
\arguments{
\item{veg_cover}{save_folder}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
...
}
