% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{predictLC}
\alias{predictLC}
\title{predictLC}
\usage{
predictLC(model, veg_predictors, meta_cols, lc_modern, save_folder = NULL)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
predictLC
}
