% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{predictLcAllClasses}
\alias{predictLcAllClasses}
\title{predictLcAllClasses}
\usage{
predictLcAllClasses(lc_veg_only, lc_all_classes, verb = 0)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
predictLcAllClasses
}
