% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{predictLcExplOnly}
\alias{predictLcExplOnly}
\title{predictLcExplOnly}
\usage{
predictLcExplOnly(model, veg_predictors, meta_cols)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
predictLcExplOnly
}
