% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{projectRegion}
\alias{projectRegion}
\title{projectRegion}
\usage{
projectRegion(lc_env_df, ID, save_folder = NULL)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
projectRegion
}
