% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{psaMap}
\alias{psaMap}
\title{psaMap}
\usage{
psaMap(veg_folder, regMap, calibMap, save_folder = NULL, plt = T)
}
\arguments{
\item{veg_folder}{character, absolute path to the folder where vegetation cover files were copied to}

\item{regMap}{sfc_MULTIPOLYGON, map for that the land cover change shall be computed}

\item{calibMap}{sfc_MULTIPOLYGON, map from which sites are selected to build the translation model from vegetation cover to landcover time series}

\item{save_folder}{character, absolute path to the directory where the results list and if plt == T the plots are saved to}

\item{plt}{boolean, whether to plot the pollen source areas, the number of samples per site, and the oldest age of the samples per site within the maps passed}
}
\value{
a list object
\item{\code{data.frame}}{veg_cover: vegetation cover in \% per taxon with Dataset_ID, Age, Intersects_regMap, Intersects_calibMap columns}
\item{\code{...}}{...: ...}
}
\description{
The function \code{psaMap} loads all vegetation cover files from the specified folder, builds pollen source areas (PSA), i.e. circular polygons around the site's centres with its pollen source radius.
The vegetation cover is filtered to all PSAs within the two maps provided (for computing past land cover and for the calibration of the model for vegetation cover to land cover translation.)
}
