% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{rdaModel}
\alias{rdaModel}
\title{rdaModel}
\usage{
rdaModel(
  lc_env_rda,
  meta_cols,
  significance_tests = F,
  p_val = 0.05,
  plt_model = F,
  width = 1200,
  height = 800,
  save_folder = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
rdaModel
}
