% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{readWhatever}
\alias{readWhatever}
\title{Read Whatever}
\usage{
readWhatever(file, skip = 0)
}
\arguments{
\item{file}{the light file to import.}

\item{skip}{number of initial lines to skip}
}
\value{
Returns a dataframe with columns
\item{\code{data.frame}}{a valid data frame if all goes well}
}
\description{
Read data
}
\details{
The \code{readWhatever} function is and example function.
}
