% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{regGrid}
\alias{regGrid}
\title{regGrid}
\usage{
regGrid(map, cell_length, save_folder = NULL, width = 1200, height = 800)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
regGrid
}
