% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandCoverChange.R
\name{vegLcModel}
\alias{vegLcModel}
\title{vegLcModel}
\usage{
vegLcModel(
  veg_modern,
  meta_cols_veg,
  lc_modern,
  meta_cols_lc,
  prefilter_taxa = F,
  tune_taxa = F,
  with_ic = F,
  save_folder = NULL
)
}
\arguments{
\item{save_folder}{}
}
\value{
...
\item{\code{...}}{...: ...}
}
\description{
vegLcModel
}
